changed character encoding of decoder
according to specification:
If nothing else is said a string is represented as ISO-8859-1
[ISO-8859-1] characters in the range $20 - $FF. All unicode strings
[UNICODE] use 16-bit unicode 2.0 (ISO/IEC 10646-1:1993, UCS-2). All
numeric strings are always encoded as ISO-8859-1. Terminated strings
are terminated with $00 if encoded with ISO-8859-1 and $00 00 if
encoded as unicode. If nothing else is said newline character is
forbidden. In ISO-8859-1 a new line is represented, when allowed, with
$0A only. Frames that allow different types of text encoding have a
text encoding description byte directly after the frame size. If
ISO-8859-1 is used this byte should be $00, if unicode is used it
should be $01.